+++ getmac/__main__.py
@@ -10,26 +10,33 @@ from . import getmac
 
 
 def main():
-    description = 'Get the MAC address of system network ' \
-                  'interfaces or remote hosts on the LAN'
-    parser = argparse.ArgumentParser('getmac', description=description)
-    parser.add_argument('--version', action='version',
-                        version='getmac %s' % getmac.__version__)
-    parser.add_argument('-d', '--debug', action='count',
-                        help='Enable debugging output. Add characters to '
-                             'increase verbosity of output, e.g. \'-dd\'.')
-    parser.add_argument('--no-network-requests', action='store_true',
-                        help='Do not send a UDP packet to refresh the ARP table')
+    parser = argparse.ArgumentParser(
+        'getmac', description='Get the MAC address of system network '
+                              'interfaces or remote hosts on the LAN')
+    parser.add_argument(
+        '--version', action='version',
+        version='getmac %s' % getmac.__version__)
+    parser.add_argument(
+        '-d', '--debug', action='count',
+        help='Enable debugging output. Add characters to '
+             'increase verbosity of output, e.g. \'-dd\'.')
+    parser.add_argument(
+        '-N', '--no-net', '--no-network-requests', action='store_true',
+        help='Do not send a UDP packet to refresh the ARP table')
 
     group = parser.add_mutually_exclusive_group(required=False)
-    group.add_argument('-i', '--interface', type=str, default=None,
-                       help='Name of a network interface on the system')
-    group.add_argument('-4', '--ip', type=str, default=None,
-                       help='IPv4 address of a remote host')
-    group.add_argument('-6', '--ip6', type=str, default=None,
-                       help='IPv6 address of a remote host')
-    group.add_argument('-n', '--hostname', type=str, default=None,
-                       help='Hostname of a remote host')
+    group.add_argument(
+        '-i', '--interface', type=str, default=None,
+        help='Name of a network interface on the system')
+    group.add_argument(
+        '-4', '--ip', type=str, default=None,
+        help='IPv4 address of a remote host')
+    group.add_argument(
+        '-6', '--ip6', type=str, default=None,
+        help='IPv6 address of a remote host')
+    group.add_argument(
+        '-n', '--hostname', type=str, default=None,
+        help='Hostname of a remote host')
 
     args = parser.parse_args()
     if args.debug: